+++ CHANGELOG.md
@@ -11,6 +11,21 @@ Note that Sockeye has checks in place to not translate with an old model that wa
 
 Each version section may have subsections for: _Added_, _Changed_, _Removed_, _Deprecated_, and _Fixed_.
 
+## [3.1.26]
+
+### Added
+
+- `--tf32 0|1` bool device (`torch.backends.cuda.matmul.allow_tf32`)
+ enabling 10-bit precision (19 bit total) transparent float32
+ acceleration. default true for backward compat with torch < 1.12.
+ allow different `--tf32` training continuation
+
+### Changed
+
+- device.init_device called by train, translate, and score
+
+- allow torch 1.12 in requirements.txt
+
 ## [3.1.25]
 
 ## Changed